timer: use absolute time instead of tick
When adding a new timer in the timer list, the TICK OS is used to
calculate timeout.

But in IRQ context, with tickless idle, the total number of elapsed
ticks is not up-to-date (as it is updated at exit of IRQ). This leads
to wrong timer expiration.

This new implementation uses the absolute time for timeout calculation.

Updated documentation about the use of the elapsed ticks (as not accurate in
IRQ context).

The publish_timeout internal function takes milliseconds as arguments instead
of ticks.

Change-Id: I23f11d13f2bdd7605c5c815916f2de4e0a0964d0
Tracked-On: https://jira.ndg.intel.com/browse/FIRE-5774
Signed-off-by: Simon Desfarges <simon.desfarges@intel.com>
Reviewed-on: https://android.intel.com/507776
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Chereau, Fabien <fabien.chereau@intel.com>
(cherry picked from commit 591b0821ad2a7dfa1505b5083292dcf0aa93ae8a)
Reviewed-on: https://android.intel.com/508523
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>