kconfig: fixed several kconfig dependency issues
KCONFIG_CONF and KCONFIG_MCONF tools were added to the HOST_TOOLS variable
that lists common tools that must be built before any targets, but the
file that defines them was never sourced.

In addition, adding both conf and mconf as equivalent dependencies led
to a race condition: we now just add one, as the two are always built
together.

Finally, both conf and mconf has a dependency towards the build directory,
which seemed to trigger a rebuild when building several targets in
parallel.

Change-Id: Ie054e0bacd09be46229d2fd6b0aa8526f6514dbc
Tracked-On: https://jira.ndg.intel.com/browse/FIRE-5964
Signed-off-by: David Corvoysier <david.corvoysier@intel.com>
Reviewed-on: https://android.intel.com/510968
Tested-by: Chevalier, Richard <richard.chevalier@intel.com>
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Chevalier, Richard <richard.chevalier@intel.com>